fix(optee): match variable type
This corrects MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
Changed type of variable to match return type.

+++ services/spd/opteed/opteed_pm.c
@@ -65,7 +65,7 @@ static int32_t opteed_cpu_off_handler(u_register_t unused)
  ******************************************************************************/
 static void opteed_cpu_suspend_handler(u_register_t max_off_pwrlvl)
 {
-	int32_t rc = 0;
+	uint64_t rc = 0U;
 	uint32_t linear_id = plat_my_core_pos();
 	optee_context_t *optee_ctx = &opteed_sp_context[linear_id];
 
@@ -87,7 +87,7 @@ static void opteed_cpu_suspend_handler(u_register_t max_off_pwrlvl)
 	 * Read the response from OPTEE. A non-zero return means that
 	 * something went wrong while communicating with OPTEE.
 	 */
-	if (rc != 0)
+	if (rc != 0U)
 		panic();
 
 	/* Update its context to reflect the state OPTEE is in */
